requirements.txt
discord.py>=2.7.0
aiohttp>=3.8.0
discord-ext-menus>=1.1.0
jishaku>=2.6.3
mcstatus>=5.0.0
wavelink>=2.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
aiosqlite>=0.22.0
pymongo>=4.0.0
motor>=3.7.1
numpy>=1.26.0
pyttsx3>=2.99
gtts>=2.5.4
langdetect>=1.0.9
PyYAML>=6.0
requests>=2.32.0
